pyproject.toml
[project]
name = "SkillScout"
version = "0.1.0"
description = "Pakistan's first data-driven career intelligence platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    'psycopg[binary]>=3.1.0',
    'znpg>=1.1.0',
    'pandas>=2.0.0',
    'numpy>=1.24.0',
    "playwright>=1.30.0",
    'requests>=2.31.0',
    'beautifulsoup4>=4.12.0',
    'python-dotenv>=1.0.0',
    'pyyaml>=6.0',
    'pytest>=7.4.0',
    "query-builder>=0.0.7",
    "dash>=3.4.0",
    "fastapi[standard]>=0.128.0",
    "uvicorn>=0.40.0",
]
